refactor(team): extract runtime snapshot run types and member lookup
Move the runtime snapshot run shape declarations and the exact-match
member lookup helpers out of TeamProvisioningRuntimeSnapshot.ts into two
sibling modules, restoring the dependency direction used by the rest of
the provisioning family (parent imports from child, never the reverse).

TeamProvisioningRuntimeSnapshotTypes.ts now declares
RuntimeAdapterRunSnapshotSource and TeamProvisioningRuntimeSnapshotRun.

TeamProvisioningRuntimeSnapshotMemberLookup.ts imports those two types
and holds the nine lookup helpers (findExactMemberRecordEntry,
findExactPersistedLaunchMember, memberIdentityKey,
findExactActiveRunMember, findExactActiveRunMemberModel,
findExactTrackedMemberSpawnStatus, findExactMapEntry,
resolveActiveRunLaneIdentity, resolveActiveRunRuntimeAdapterEvidence)
plus their three private sub-helpers, which stay module-private.

The four files that referenced the two types now import them from the
types module. PersistedRuntimeMemberLike stays in the original file
because eleven production files and two tests import it from there.

No behaviour change: the moved code is byte-identical apart from the
added export keywords, and TeamCreateRequest is dropped from the
original file's shared-types import because it now has no remaining use
there. Existing tests pass unchanged; the only test edit is the import
repoint.
