+++ src/main/services/team/provisioning/TeamProvisioningLiveRuntimeMetadataPortsFactory.ts
@@ -1,12 +1,14 @@
 import {
   buildLiveTeamAgentRuntimeMetadata,
   type PersistedRuntimeMemberLike,
-  type RuntimeAdapterRunSnapshotSource,
-  type TeamProvisioningRuntimeSnapshotRun,
 } from './TeamProvisioningRuntimeSnapshot';
 
 import type { LiveTeamAgentRuntimeMetadata } from './TeamProvisioningRuntimeMetadataPolicy';
 import type { TeamProvisioningLiveRuntimeMetadataCachePort } from './TeamProvisioningRuntimeSnapshotCache';
+import type {
+  RuntimeAdapterRunSnapshotSource,
+  TeamProvisioningRuntimeSnapshotRun,
+} from './TeamProvisioningRuntimeSnapshotTypes';
 
 type BuildLiveTeamAgentRuntimeMetadataParams = Parameters<
   typeof buildLiveTeamAgentRuntimeMetadata
